refactor: move cost models to common
