+++ common/src/lib.rs
@@ -4,6 +4,7 @@
 pub mod address;
 pub mod allocations;
 pub mod attestations;
+pub mod cost_model;
 pub mod escrow_accounts;
 pub mod graphql;
 pub mod indexer_service;
